tgbot/pyproject.toml
[project]
name = "tgbot-usfeu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "aiogram>=3.23.0",
    "aiosqlite>=0.22.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "ruff>=0.14.8",
]

[tool.ruff]
target-version = "py313"
line-length = 100
extend-exclude = ["app/migration/"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
ignore = ["E501", "B008", "S101","B904"]

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.mypy]
# Основные опции
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
show_error_codes = true

# Проверки
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports=true

# Строгость
strict_equality = true
strict_optional = true

# Кеширование
cache_dir = ".mypy_cache"

# Формат вывода
show_column_numbers = true
pretty = true

# Игнорирование файлов
exclude = [
    "tests/",
    "test_",
    ".*_test.py",
    "migrations/",
    "__pycache__/",
    ".*\\.pyc",
    "setup.py",
    "build/",
    "dist/",
    ".venv/",
    "venv/",
]

# Или более тонкое управление
[tool.mypy-plugins]
plugins = [
    "pydantic.mypy",
]
